# perf(web): optimize encodeInto()