--- a/core/Cargo.toml
+++ b/core/Cargo.toml
@@ -33,7 +33,7 @@ serde_json = { version = "1.0.79", features = ["preserve_order"] }
 serde_v8 = { version = "0.61.0", path = "../serde_v8" }
 sourcemap = "=6.0.1"
 url = { version = "2.3.1", features = ["serde", "expose_internals"] }
-v8 = { version = "0.49.0", default-features = false }
+v8 = { version = "0.49.0", default-features = false, path = "../../rusty_v8" }
 
 [[example]]
 name = "http_bench_json_ops"