--- a/ext/web/08_text_encoding.js
+++ b/ext/web/08_text_encoding.js
@@ -199,9 +199,15 @@
         context: "Argument 2",
         allowShared: true,
       });
-      return ops.op_encoding_encode_into(source, destination);
+      ops.op_encoding_encode_into(source, destination, encodeIntoBuf);
+      return {
+        read: encodeIntoBuf[0],
+        written: encodeIntoBuf[1],
+      };
     }
   }
+
+  const encodeIntoBuf = new Uint32Array(2);
 
   webidl.configurePrototype(TextEncoder);
   const TextEncoderPrototype = TextEncoder.prototype;