--- a/ext/web/08_text_encoding.js
+++ b/ext/web/08_text_encoding.js
@@ -25,6 +25,7 @@
     TypedArrayPrototypeSubarray,
     TypedArrayPrototypeSlice,
     Uint8Array,
+    Uint32Array,
   } = window.__bootstrap.primordials;
 
   class TextDecoder {