Towards https://github.com/denoland/deno/issues/15895

- [x] Use a preallocated Uint32Array result buffer. Much more efficient then serializing an object.

- [x] Use `v8::String::write_utf8` directly on the backing store.

- [x] Use `v8::ArrayBuffer::data` instead of going through the backing store abstraction. See https://v8.github.io/api/head/classv8_1_1ArrayBuffer.html
  > More efficient shortcut for [GetBackingStore()](https://v8.github.io/api/head/classv8_1_1ArrayBuffer.html#aaa29e281f94990cbf0a5484140fffb50)->Data(). The returned pointer is valid as long as the [ArrayBuffer](https://v8.github.io/api/head/classv8_1_1ArrayBuffer.html) is alive.

  https://github.com/denoland/rusty_v8/pull/1068
  
5.4x faster encodeInto()

```
# This patch
target/release/deno run -A --unstable cli/bench/encode_into.js
time 78 ms rate 12820512
time 66 ms rate 15151515
time 65 ms rate 15384615
time 66 ms rate 15151515
```

```
# main
deno run -A --unstable cli/bench/encode_into.js
time 362 ms rate 2762430
time 340 ms rate 2941176
time 340 ms rate 2941176
time 339 ms rate 2949852
```