@Phosra 2022-09-16
This this same optimisation be applicable to `encode()` - the difference being who allocates the return `Uint8Array`?