@billywhizz 2022-09-17
i benchmarked this change against current implementation: https://github.com/denoland/deno/issues/15895#issuecomment-1249977244